Fix for Kyoto capinet error.
OS X Yosemite comes with a C++11 compiler. To be able to compile the
standard had to be specified to the compiler (build.rules). Now the
headeres can be found, but there is an error in Kyoto library. It is
corrected.

Two warnings remained int he code. But I do not want to fix it right now.